requirements.txt
# Core dependencies for RAG system
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
chromadb>=0.4.22
faiss-cpu>=1.9.0
sentence-transformers>=2.2.2
transformers>=4.36.2
torch>=2.1.2

# Document processing
pypdf2>=3.0.1
python-docx>=1.1.0
unstructured>=0.11.6
pymupdf>=1.23.8

# Web interface
streamlit>=1.29.0
streamlit-chat>=0.1.1

# Utilities
python-dotenv>=1.0.0
numpy>=1.24.3
pandas>=2.1.4
tqdm>=4.66.1
requests>=2.31.0
scikit-learn>=1.3.0
plotly>=5.17.0

# Optional: For advanced retrieval
rank-bm25>=0.2.2
